requirements.txt
numpy 
scipy
pandas
matplotlib
cmasher
scikit-image
tqdm
numpy-stl